cleanup restructure
