+++ embc-app/ClientApp/src/app/app-routing.module.ts
@@ -24,16 +24,16 @@ import { TaskNumberMakerComponent } from './provincial-admin/components/task-num
 import { VolunteerMakerComponent } from './volunteer-maker/volunteer-maker.component';
 import { RegistrationSummaryFullComponent } from './registration-summary-full/registration-summary-full.component';
 import { SessionExpiredComponent } from './session-expired/session-expired.component';
-import { UsefulInformationPageComponent } from './pages/useful-information-page/useful-information-page.component';
-import { LocalAuthorityRegistrationsPageComponent } from './pages/local-authority/local-authority-registrations-page/local-authority-registrations-page.component';
+import { UsefulInformationPageComponent } from './volunteer/pages/useful-information-page/useful-information-page.component';
+import { LocalAuthorityRegistrationsPageComponent } from './local-authority/local-authority-registrations-page/local-authority-registrations-page.component';
 import { ProvincialAdminOrganizationsPageComponent } from './provincial-admin/pages/organizations/provincial-admin-organizations-page.component';
 import { ProvincialAdminTaskNumbersPageComponent } from './provincial-admin/pages/task-numbers/provincial-admin-task-numbers-page.component';
 import { ProvincialAdminRegistrationsPageComponent } from './provincial-admin/pages/registrations/provincial-admin-registrations-page.component';
-import { LocalAuthorityVolunteersPageComponent } from './pages/local-authority/local-authority-volunteers-page/local-authority-volunteers-page.component';
+import { LocalAuthorityVolunteersPageComponent } from './local-authority/local-authority-volunteers-page/local-authority-volunteers-page.component';
 import { ReferralMakerComponent } from './volunteer/pages/referral-maker/referral-maker.component';
 import { ReferralViewComponent } from './volunteer/pages/referral-view/referral-view.component';
 import { ProvincialAdminVolunteersOrganizationPageComponent } from './provincial-admin/pages/volunteers-organization/provincial-admin-volunteers-organization-page.component';
-import { VolunteerRegistrationsPageComponent } from './pages/volunteer-registrations-page/volunteer-registrations-page.component';
+import { VolunteerRegistrationsPageComponent } from './volunteer/pages/volunteer-registrations-page/volunteer-registrations-page.component';
 
 /*
   /
